+++ components/layout/SidebarMain.tsx
@@ -2,11 +2,9 @@
 
 import { useEffect, useState } from 'react';
 import { getCookie } from 'cookies-next';
-import { usePathname } from 'next/navigation';
 import Link from 'next/link';
 import { ChevronLeft } from 'lucide-react';
 import { AgentSelector } from '@/components/layout/agent-selector';
-import { ChatHistory } from '@/components/layout/chat-history';
 import { NavMain } from '@/components/layout/NavMain';
 import { NavUser } from '@/components/layout/NavUser';
 import { useUser } from '@/components/interactive/useUser';
@@ -20,8 +18,176 @@ import {
   SidebarMenu,
   SidebarMenuButton,
   SidebarMenuItem,
+  SidebarGroup,
+  SidebarGroupLabel,
   useSidebar,
 } from '@/components/ui/sidebar';
+('use client');
+
+import { Badge } from '@/components/ui/badge';
+import { Tooltip, TooltipContent, TooltipTrigger } from '@/components/ui/tooltip';
+import { getTimeDifference } from '@/components/conversation/activity';
+import { cn } from '@/lib/utils';
+import { DotsHorizontalIcon } from '@radix-ui/react-icons';
+import dayjs from 'dayjs';
+import { usePathname, useRouter } from 'next/navigation';
+import { useContext } from 'react';
+import { Command, CommandInput, CommandItem, CommandList } from '@/components/ui/command';
+import { Dialog, DialogClose, DialogContent, DialogTrigger } from '@/components/ui/dialog';
+import { Conversation, useConversations } from '@/components/interactive/useConversation';
+import { InteractiveConfigContext } from '@/components/interactive/InteractiveConfigContext';
+
+export function ChatHistory() {
+  const state = useContext(InteractiveConfigContext);
+  const { data: conversationData, isLoading } = useConversations();
+  const pathname = usePathname();
+  const router = useRouter();
+
+  const isActive = (conversationId: string) => pathname.includes('chat') && pathname.includes(conversationId);
+
+  const handleOpenConversation = ({ conversationId }: { conversationId: string | number }) => {
+    router.push(`/chat/${conversationId}`);
+
+    state?.mutate?.((oldState) => ({
+      ...oldState,
+      overrides: { ...oldState.overrides, conversation: conversationId },
+    }));
+  };
+
+  if (!conversationData || !conversationData.length || isLoading) return null;
+  const groupedConversations = groupConversations(conversationData.filter((conversation) => conversation.name !== '-'));
+
+  return (
+    <SidebarGroup className='group-data-[collapsible=icon]:hidden'>
+      {Object.entries(groupedConversations).map(([label, conversations]) => (
+        <div key={label}>
+          <SidebarGroupLabel>{label}</SidebarGroupLabel>
+          <SidebarMenu className='ml-1'>
+            {conversations.map((conversation) => (
+              <SidebarMenuItem key={conversation.id}>
+                <Tooltip>
+                  <TooltipTrigger asChild>
+                    <SidebarMenuButton
+                      side='left'
+                      onClick={() => handleOpenConversation({ conversationId: conversation.id })}
+                      className={cn(
+                        'flex items-center justify-between w-full transition-colors',
+                        isActive(conversation.id) && 'bg-sidebar-accent text-sidebar-accent-foreground font-medium',
+                      )}
+                    >
+                      <span className='truncate'>{conversation.name}</span>
+                      {conversation.hasNotifications && (
+                        <Badge
+                          variant='default'
+                          className={cn(
+                            'ml-2',
+                            isActive(conversation.id)
+                              ? 'bg-sidebar-accent-foreground/10 text-sidebar-accent-foreground'
+                              : 'bg-primary/10 text-primary',
+                          )}
+                        >
+                          New
+                        </Badge>
+                      )}
+                    </SidebarMenuButton>
+                  </TooltipTrigger>
+                  <TooltipContent side='right'>
+                    <div>{conversation.name}</div>
+                    {label === 'Today' ? (
+                      <div>
+                        Updated: {getTimeDifference(dayjs().format('YYYY-MM-DDTHH:mm:ssZ'), conversation.updatedAt)} ago
+                      </div>
+                    ) : (
+                      <div>Updated: {dayjs(conversation.updatedAt).format('MMM DD YYYY')}</div>
+                    )}
+                    {conversation.attachmentCount > 0 && <div>Attachments: {conversation.attachmentCount}</div>}
+                  </TooltipContent>
+                </Tooltip>
+              </SidebarMenuItem>
+            ))}
+          </SidebarMenu>
+        </div>
+      ))}
+      <SidebarMenu>
+        <SidebarMenuItem>
+          {conversationData && conversationData?.length > 10 && (
+            <ChatSearch {...{ conversationData, handleOpenConversation }}>
+              <SidebarMenuItem>
+                <SidebarMenuButton className='text-sidebar-foreground/70' side='left'>
+                  <DotsHorizontalIcon className='text-sidebar-foreground/70' />
+                  <span>More</span>
+                </SidebarMenuButton>
+              </SidebarMenuItem>
+            </ChatSearch>
+          )}
+        </SidebarMenuItem>
+      </SidebarMenu>
+    </SidebarGroup>
+  );
+}
+
+function ChatSearch({
+  conversationData,
+  handleOpenConversation,
+  children,
+}: {
+  conversationData: any[];
+  handleOpenConversation: ({ conversationId }: { conversationId: string | number }) => void;
+  children: React.ReactNode;
+}) {
+  return (
+    <Dialog>
+      <DialogTrigger asChild>{children}</DialogTrigger>
+      <DialogContent className='p-0 overflow-hidden shadow-lg'>
+        <Command className='[&_[cmdk-group-heading]]:px-2 [&_[cmdk-group-heading]]:font-medium [&_[cmdk-group-heading]]:text-muted-foreground [&_[cmdk-group]:not([hidden])_~[cmdk-group]]:pt-0 [&_[cmdk-group]]:px-2 [&_[cmdk-input-wrapper]_svg]:h-5 [&_[cmdk-input-wrapper]_svg]:w-5 [&_[cmdk-input]]:h-12 [&_[cmdk-item]]:px-2 [&_[cmdk-item]]:py-3 [&_[cmdk-item]_svg]:h-5 [&_[cmdk-item]_svg]:w-5'>
+          <CommandInput placeholder='Search Conversations...' />
+          <CommandList>
+            {conversationData.map((conversation) => (
+              <CommandItem asChild key={conversation.id}>
+                <DialogClose className='w-full' onClick={() => handleOpenConversation({ conversationId: conversation.id })}>
+                  <span className='px-2'>{conversation.name}</span>
+                </DialogClose>
+              </CommandItem>
+            ))}
+          </CommandList>
+        </Command>
+      </DialogContent>
+    </Dialog>
+  );
+}
+
+function groupConversations(conversations: Conversation[]) {
+  const today = new Date();
+  const yesterday = new Date();
+  yesterday.setDate(today.getDate() - 1);
+
+  const isToday = (date: string) => new Date(date).toDateString() === today.toDateString();
+  const isYesterday = (date: string) => new Date(date).toDateString() === yesterday.toDateString();
+  const isPastWeek = (date: string) => {
+    const d = new Date(date);
+    const weekAgo = new Date();
+    weekAgo.setDate(today.getDate() - 7);
+    return d > weekAgo && d < yesterday;
+  };
+
+  const groups = conversations.slice(0, 7).reduce(
+    (groups: { [key: string]: Conversation[] }, conversation: Conversation) => {
+      if (isToday(conversation.updatedAt)) {
+        groups['Today'].push(conversation);
+      } else if (isYesterday(conversation.updatedAt)) {
+        groups['Yesterday'].push(conversation);
+      } else if (isPastWeek(conversation.updatedAt)) {
+        groups['Past Week'].push(conversation);
+      } else {
+        groups['Older'].push(conversation);
+      }
+      return groups;
+    },
+    { Today: [], Yesterday: [], 'Past Week': [], Older: [] },
+  );
+
+  return Object.fromEntries(Object.entries(groups).filter(([_, conversations]) => conversations.length > 0));
+}
 
 export function ToggleSidebar({ side }: { side: 'left' | 'right' }) {
   const { toggleSidebar } = useSidebar(side);
